requirements.txt
charset-normalizer==3.4.0
contourpy==1.3.1
cycler==0.12.1
efficientnet-pytorch==0.7.1
filelock==3.13.1
fonttools==4.55.0
fsspec==2024.2.0
huggingface-hub==0.26.3
idna==3.10
jinja2==3.1.3
kiwisolver==1.4.7
markupsafe==2.1.5
matplotlib==3.9.3
mpmath==1.3.0
munch==4.0.0
networkx==3.2.1
numpy==1.26.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
opencv-python==4.10.0.84
packaging==24.2
pillow==10.2.0
pretrainedmodels==0.7.4
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pyyaml==6.0.2

Flask~=3.1.0
torch~=2.5.1
matplotlib~=3.9.3
torchvision~=0.20.1
pillow~=11.0.0
segmentation-models-pytorch~=0.3.4